backend/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
numpy
pydantic
python-multipart
git+https://github.com/jacobgil/pytorch-grad-cam.git

